feat: 식단 데이터 로컬 캐시로 성능 개선
