+++ lib/notice_page/entity/GeneralNotice.dart
