+++ lib/notice_page/entity/Notice.dart
